+++ src/main/java/com/skuniv/fuwarilog/controller/DiaryController.java
@@ -2,12 +2,10 @@ package com.skuniv.fuwarilog.controller;
 
 import com.skuniv.fuwarilog.config.exception.BadRequestException;
 import com.skuniv.fuwarilog.config.exception.ErrorResponseStatus;
-import com.skuniv.fuwarilog.domain.Diary;
 import com.skuniv.fuwarilog.domain.DiaryContent;
-import com.skuniv.fuwarilog.dto.DiaryContentRequest;
-import com.skuniv.fuwarilog.dto.DiaryListResponse;
-import com.skuniv.fuwarilog.dto.DiaryResponse;
-import com.skuniv.fuwarilog.dto.TripResponse;
+import com.skuniv.fuwarilog.dto.DiaryContent.DiaryContentRequest;
+import com.skuniv.fuwarilog.dto.DiaryList.DiaryListResponse;
+import com.skuniv.fuwarilog.dto.Trip.TripResponse;
 import com.skuniv.fuwarilog.security.jwt.JwtTokenProvider;
 import com.skuniv.fuwarilog.service.DiaryService;
 import io.swagger.v3.oas.annotations.Operation;
