+++ src/main/java/com/skuniv/fuwarilog/dto/Auth/AuthResponse.java
@@ -1,11 +1,9 @@
-package com.skuniv.fuwarilog.dto;
+package com.skuniv.fuwarilog.dto.Auth;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.time.LocalDateTime;
 
 public class AuthResponse {
     @Getter
