+++ src/main/java/com/skuniv/fuwarilog/dto/DiaryContent/DiaryContentRequest.java
@@ -1,11 +1,8 @@
-package com.skuniv.fuwarilog.dto;
+package com.skuniv.fuwarilog.dto.DiaryContent;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.time.LocalDate;
 
 public class DiaryContentRequest {
