+++ src/main/java/com/skuniv/fuwarilog/dto/DiaryList/DiaryListResponse.java
@@ -1,4 +1,4 @@
-package com.skuniv.fuwarilog.dto;
+package com.skuniv.fuwarilog.dto.DiaryList;
 
 import com.skuniv.fuwarilog.domain.DiaryList;
 import io.swagger.v3.oas.annotations.media.Schema;
