+++ src/main/java/com/skuniv/fuwarilog/dto/ExchangeRate/ExchangeRateRequest.java
@@ -1,10 +1,9 @@
-package com.skuniv.fuwarilog.dto;
+package com.skuniv.fuwarilog.dto.ExchangeRate;
 
 import com.skuniv.fuwarilog.domain.Currency;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 
-import java.sql.Timestamp;
 import java.time.LocalDateTime;
 
 public class ExchangeRateRequest {
