+++ src/main/java/com/skuniv/fuwarilog/dto/ExchangeRate/ExchangeResponse.java
@@ -1,0 +1,4 @@
+package com.skuniv.fuwarilog.dto.ExchangeRate;
+
+public class ExchangeResponse {
+}
