--- src/main/java/com/skuniv/fuwarilog/dto/ExchangeResponse.java
@@ -1,4 +1,0 @@
-package com.skuniv.fuwarilog.dto;
-
-public class ExchangeResponse {
-}
