+++ src/main/java/com/skuniv/fuwarilog/dto/Location/LocationResponse.java
@@ -1,4 +1,4 @@
-package com.skuniv.fuwarilog.dto;
+package com.skuniv.fuwarilog.dto.Location;
 
 import com.skuniv.fuwarilog.domain.Location;
 import io.swagger.v3.oas.annotations.media.Schema;
