+++ src/main/java/com/skuniv/fuwarilog/dto/VisitedRouteDocument/VisitedRouteDocumentResponse.java
@@ -1,10 +1,4 @@
-package com.skuniv.fuwarilog.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
+package com.skuniv.fuwarilog.dto.VisitedRouteDocument;
 
 public class VisitedRouteDocumentResponse {
     public record RouteDTO (String distanceText, String durationText, String lat, String lon) {};
