+++ src/main/java/com/skuniv/fuwarilog/service/DiaryService.java
@@ -5,10 +5,10 @@ import com.skuniv.fuwarilog.config.exception.ErrorResponseStatus;
 import com.skuniv.fuwarilog.domain.DiaryContent;
 import com.skuniv.fuwarilog.domain.DiaryList;
 import com.skuniv.fuwarilog.domain.Trip;
-import com.skuniv.fuwarilog.dto.DiaryContentRequest;
-import com.skuniv.fuwarilog.dto.DiaryListResponse;
-import com.skuniv.fuwarilog.dto.DiaryResponse;
-import com.skuniv.fuwarilog.dto.TripResponse;
+import com.skuniv.fuwarilog.dto.DiaryContent.DiaryContentRequest;
+import com.skuniv.fuwarilog.dto.DiaryList.DiaryListResponse;
+import com.skuniv.fuwarilog.dto.Diary.DiaryResponse;
+import com.skuniv.fuwarilog.dto.Trip.TripResponse;
 import com.skuniv.fuwarilog.repository.DiaryContentRepository;
 import com.skuniv.fuwarilog.repository.DiaryListRepository;
 import com.skuniv.fuwarilog.repository.TripRepository;
