+++ src/main/java/com/skuniv/fuwarilog/service/LocationService.java
@@ -3,10 +3,10 @@ package com.skuniv.fuwarilog.service;
 import com.skuniv.fuwarilog.config.exception.BadRequestException;
 import com.skuniv.fuwarilog.config.exception.ErrorResponseStatus;
 import com.skuniv.fuwarilog.domain.*;
-import com.skuniv.fuwarilog.dto.LocationRequest;
-import com.skuniv.fuwarilog.dto.LocationResponse;
-import com.skuniv.fuwarilog.dto.VisitedRouteDocumentRequest;
-import com.skuniv.fuwarilog.dto.VisitedRouteDocumentResponse;
+import com.skuniv.fuwarilog.dto.Location.LocationRequest;
+import com.skuniv.fuwarilog.dto.Location.LocationResponse;
+import com.skuniv.fuwarilog.dto.VisitedRouteDocument.VisitedRouteDocumentRequest;
+import com.skuniv.fuwarilog.dto.VisitedRouteDocument.VisitedRouteDocumentResponse;
 import com.skuniv.fuwarilog.repository.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
