+++ src/main/java/com/skuniv/fuwarilog/service/TripService.java
@@ -6,8 +6,8 @@ import com.skuniv.fuwarilog.domain.Diary;
 import com.skuniv.fuwarilog.domain.DiaryList;
 import com.skuniv.fuwarilog.domain.Trip;
 import com.skuniv.fuwarilog.domain.User;
-import com.skuniv.fuwarilog.dto.TripRequest;
-import com.skuniv.fuwarilog.dto.TripResponse;
+import com.skuniv.fuwarilog.dto.Trip.TripRequest;
+import com.skuniv.fuwarilog.dto.Trip.TripResponse;
 import com.skuniv.fuwarilog.repository.DiaryListRepository;
 import com.skuniv.fuwarilog.repository.DiaryRepository;
 import com.skuniv.fuwarilog.repository.TripRepository;
