+++ src/main/java/com/skuniv/fuwarilog/service/UserService.java
@@ -3,8 +3,8 @@ package com.skuniv.fuwarilog.service;
 import com.skuniv.fuwarilog.config.exception.BadRequestException;
 import com.skuniv.fuwarilog.config.exception.ErrorResponseStatus;
 import com.skuniv.fuwarilog.domain.User;
-import com.skuniv.fuwarilog.dto.UserRequest;
-import com.skuniv.fuwarilog.dto.UserResponse;
+import com.skuniv.fuwarilog.dto.User.UserRequest;
+import com.skuniv.fuwarilog.dto.User.UserResponse;
 import com.skuniv.fuwarilog.repository.UserRepository;
 import com.skuniv.fuwarilog.security.jwt.JwtTokenProvider;
 import jakarta.transaction.Transactional;
